ENH: Make df as Crosstab instance variable.
It would be better if df is Crosstab instance variable so that pandas
DataFrame is available for use on objects where Crosstab is passed as
parameters. It would be even better if Crosstab can be a child class
of pandas DataFrame, makes the code much cleaner. But I still don't know
how to do this. I have to figure this out. Meanwhile an issue #22 is raise
for this.